requirements.txt
accelerate==0.32.1
datasets==2.20.0
mosaicml-streaming==0.7.5
ninja==1.11.1.1
numpy==1.26.4
pandas==2.2.1
protobuf==4.25.3
python-dateutil==2.9.0
regex==2023.12.25
sentencepiece==0.2.0
tiktoken==0.7.0
torch==2.4.0
tqdm==4.66.5
transformers==4.44.2
triton==3.0.0
wandb==0.16.6
zstandard==0.23.0
zstd==1.5.5.1

flash-attn==2.6.1, --config-settings=--global-option="--no-build-isolation" 
rotary-emb @ git+https://github.com/Dao-AILab/flash-attention.git@9356a1c0389660d7e231ff3163c1ac17d9e3824a#subdirectory=csrc/rotary
